requirements.txt
Django==3.0.2
django-compat==1.0.15
django-crispy-forms==1.8.1
djangorestframework==3.11.0
gitdb2==2.0.6
GitPython==3.0.5
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
matplotlib==3.1.2
numpy==1.18.1
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
pandas==0.25.3
Pillow==7.0.0
pipenv==2018.11.26
requests==2.22.0
scipy==1.4.1
seaborn==0.9.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
tabulate==0.8.6
terminado==0.8.3
testpath==0.4.4
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu 
torchvision==0.6.0+cpu
